+++ driver_management/Driver.java
@@ -22,8 +22,8 @@ public class Driver {
     this.vehicle = vehicle;
   }
 
-  public Integer driveDistance(int distance) {
-    int timeRequired = this.vehicle.driveDistance(distance);
+  public double driveDistance(double distance) {
+    double timeRequired = this.vehicle.driveDistance(distance);
     return timeRequired;
   }
